+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+Unreleased
+----------
+- Fixed: Equal signs in extension tag attributes are no longer confused with name-value separator in arguments. (#128)
+
 v0.55.7
 -------
 - Fixed a bug in ``plain_text``. (#126)